+++ test/node/unit/iam.js
@@ -613,26 +613,26 @@ describe('corbel IAM module', function() {
     describe('User devices', function() {
 
         it('Register my device', function() {
-            corbelDriver.iam.user().registerMyDevice('Device data');
+            corbelDriver.iam.user().registerMyDevice('deviceId', 'Device data');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device/deviceId');
             expect(callRequestParam.method).to.be.equal('PUT');
         });
 
         it('Register device', function() {
-            corbelDriver.iam.user('userId').registerDevice('Device data');
+            corbelDriver.iam.user('userId').registerDevice('deviceId', 'Device data');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/devices');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/device/deviceId');
             expect(callRequestParam.method).to.be.equal('PUT');
         });
 
         it('Register device using me', function() {
-            corbelDriver.iam.user('me').registerDevice('Device data');
+            corbelDriver.iam.user('me').registerDevice('deviceId', 'Device data');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device/deviceId');
             expect(callRequestParam.method).to.be.equal('PUT');
         });
 
@@ -640,7 +640,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user('userId').getDevice('deviceId');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/devices/deviceId');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/device/deviceId');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
@@ -648,7 +648,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user('me').getDevice('deviceId');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices/deviceId');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device/deviceId');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
@@ -656,7 +656,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user().getMyDevice('deviceId');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices/deviceId');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device/deviceId');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
@@ -664,7 +664,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user('userId').getDevices();
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/devices');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/device');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
@@ -679,7 +679,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user('userId').getDevices(params);
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/devices?api:query=%7B%22%24eq%22%3A%7B%22test%22%3A3%7D%7D');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/userId/device?api:query=%7B%22%24eq%22%3A%7B%22test%22%3A3%7D%7D');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
@@ -687,7 +687,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user('me').getDevices();
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
@@ -696,14 +696,14 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user().getMyDevices();
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
         it('Get my device', function() {
             corbelDriver.iam.user('me').getDevice('123');
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices/123');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device/123');
             expect(callRequestParam.method).to.be.equal('GET');
         });
 
@@ -711,7 +711,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user('me').deleteDevice('deviceId');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/devices/deviceId');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/me/device/deviceId');
             expect(callRequestParam.method).to.be.equal('DELETE');
         });
 
@@ -719,7 +719,7 @@ describe('corbel IAM module', function() {
             corbelDriver.iam.user('123').deleteDevice('deviceId');
 
             var callRequestParam = corbel.request.send.firstCall.args[0];
-            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/123/devices/deviceId');
+            expect(callRequestParam.url).to.be.equal(IAM_END_POINT + 'user/123/device/deviceId');
             expect(callRequestParam.method).to.be.equal('DELETE');
         });